+++ src/templates/react-app/react/base/src/components/types/StringCell.js
@@ -1,0 +1,7 @@
+function renderValue(value) {
+  return !value || value === '' ? '--' : value;
+}
+
+export function StringCell({ value }) {
+  return <div> {renderValue(value)} </div>;
+}